Add InvertedRigid for Matrix4f + a few other minor math changes

+++ Source/Common/Math/Math.h
@@ -2,6 +2,6 @@
 #include <numbers>
 namespace MagicRed {
 inline float deg2rad(float degrees) {
-    return degrees * (std::numbers::pi / 180.0f);
+    return static_cast<float>(degrees * (std::numbers::pi / 180.0f));
 }
 }
